refactor: move duration/name formatters into utils_service
humanize_duration, format_wall_clock, format_mmss, sanitize_name and
MAX_NAME_LEN now live in service/utils_service.rs so gather and break
share a single source instead of one importing from the other.
