+++ src/embeds/gather_embed.rs
@@ -1,3 +1,4 @@
+use crate::service::utils_service::{format_mmss, format_wall_clock, humanize_duration, MAX_NAME_LEN};
 use serenity::all::{
     ButtonStyle, Color, CreateActionRow, CreateButton, CreateEmbed, CreateEmbedFooter,
 };
@@ -10,7 +11,6 @@ pub const BTN_FORCE_START: &str = "gather_force_start";
 pub const BTN_TOGGLE_SILENT: &str = "gather_toggle_silent";
 
 pub const GRACE_PERIOD: Duration = Duration::from_secs(60);
-pub const MAX_NAME_LEN: usize = 21;
 
 /// One row in the gathering check-in table — already resolved to a display
 /// name so the embed module never has to touch the serenity cache.
@@ -254,54 +254,3 @@ pub fn gather_buttons(disabled: bool, silent: bool) -> Vec<CreateActionRow> {
     ])]
 }
 
-pub fn humanize_duration(d: Duration) -> String {
-    let total = d.as_secs();
-    if total == 0 {
-        return "0 seconds".to_string();
-    }
-    let h = total / 3600;
-    let m = (total % 3600) / 60;
-    let s = total % 60;
-    let mut parts: Vec<String> = Vec::new();
-    if h > 0 {
-        parts.push(format!("{} {}", h, if h == 1 { "hour" } else { "hours" }));
-    }
-    if m > 0 {
-        parts.push(format!("{} {}", m, if m == 1 { "minute" } else { "minutes" }));
-    }
-    if s > 0 {
-        parts.push(format!("{} {}", s, if s == 1 { "second" } else { "seconds" }));
-    }
-    parts.join(" ")
-}
-
-pub fn format_wall_clock(t: OffsetDateTime) -> String {
-    format!("{:02}:{:02}:{:02}", t.hour(), t.minute(), t.second())
-}
-
-pub fn format_mmss(d: Duration) -> String {
-    let total = d.as_secs();
-    let m = total / 60;
-    let s = total % 60;
-    format!("{:02}:{:02}", m, s)
-}
-
-/// Replace emoji grapheme clusters with their `:shortcode:` then truncate to
-/// `MAX_NAME_LEN` so the table stays aligned in Discord's monospace font.
-pub fn sanitize_name(name: &str) -> String {
-    use unicode_segmentation::UnicodeSegmentation;
-
-    let mut out = String::new();
-    for g in name.graphemes(true) {
-        if let Some(emoji) = emojis::get(g) {
-            let label = emoji.shortcode().unwrap_or(emoji.name());
-            out.push(':');
-            out.push_str(label.trim_matches(':'));
-            out.push(':');
-        } else {
-            out.push_str(g);
-        }
-    }
-
-    out.chars().take(MAX_NAME_LEN).collect()
-}
